Move a bunch of stuff out of functions.php to functionality-specific files
